+++ src/Chunk.php
@@ -75,48 +75,6 @@ final class Chunk implements IteratorAggregate
         $this->lines = $lines;
     }
 
-    /**
-     * @deprecated Use start() instead
-     */
-    public function getStart(): int
-    {
-        return $this->start;
-    }
-
-    /**
-     * @deprecated Use startRange() instead
-     */
-    public function getStartRange(): int
-    {
-        return $this->startRange;
-    }
-
-    /**
-     * @deprecated Use end() instead
-     */
-    public function getEnd(): int
-    {
-        return $this->end;
-    }
-
-    /**
-     * @deprecated Use endRange() instead
-     */
-    public function getEndRange(): int
-    {
-        return $this->endRange;
-    }
-
-    /**
-     * @psalm-return list<Line>
-     *
-     * @deprecated Use lines() instead
-     */
-    public function getLines(): array
-    {
-        return $this->lines;
-    }
-
     public function getIterator(): Traversable
     {
         return new ArrayIterator($this->lines);